Extract mobile notebook shell UI module
